+++ src/main/java/com/proyecto/mc_cotizacion/dto/response/QuotationResponse.java
@@ -3,6 +3,7 @@ package com.proyecto.mc_cotizacion.dto.response;
 import com.proyecto.mc_cotizacion.entity.QuotationStatus;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -16,7 +17,7 @@ public class QuotationResponse {
     private String numberQuotation;
     private String client;
     private Date dateQuotation;
-    private Float totalAmount;
+    private BigDecimal totalAmount;
     private QuotationStatus status;
 
     @Singular
